moving classes tests
